zeno_api/pyproject.toml
[tool.poetry]
name = "zeno-api"
version = "0.1.0"
description = ""
authors = ["Zeno"]
readme = "README.md"
packages = [{include = "zeno_api"}]

[tool.poetry.dependencies]
python = "^3.11"
pycognito = "^2023.5.0"
python-dotenv = "^1.0.0"
sphinx = "^7.1.2"
sphinx-rtd-theme = "^1.2.2"
ipykernel = "^6.25.1"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
pyright = "^1.1.320"
pytest = "^7.3.2"
ruff = "^0.0.282"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["zeno_backend"]

[tool.ruff]
select = ["E", "F", "I", "PD", "N", "D", "T20"]

[tool.ruff.pydocstyle]
convention = "google"
